functions/upload_image/requirements.txt
wand==0.4.5
google-cloud-storage==1.13.0